Rename has_player_color module (too)
