+++ client-base/src/plugin/has_team.rs
